requirements.txt
torch>=2.1
torchaudio>=2.1
transformers>=4.36
scipy>=1.11
demucs>=4.0
typer[all]>=0.9
pyyaml>=6.0
rich>=13.0
soundfile>=0.12
